ajout fonctionnalité help + correction documentation et indentation (#31)
* ajout fonctionnalité help + correction documentation

* Update cron-message-saint.ts

* Update embed-roll.ts

* Update message.saint.ts

+++ src/services/embed-roll.js
@@ -64,7 +64,10 @@ let EmbedRoll = class EmbedRoll {
         setTimeout(() => {
             const channel = message.guild.channels.cache
                 .find((channel) => channel.name.startsWith("appel"));
-            channel.delete();
+            if (channel) {
+                channel.delete();
+            }
+            return;
         }, 10000);
         return Promise.reject();
     }
